// web/src/app/api/generated/src/models/Question.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Motion Ecosystem API
 * API for participant and specialist workflows
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface Question
 */
export interface Question {
  /**
   *
   * @type {string}
   * @memberof Question
   */
  code?: string;
  /**
   *
   * @type {boolean}
   * @memberof Question
   */
  required?: boolean;
  /**
   *
   * @type {string}
   * @memberof Question
   */
  section?: string;
  /**
   *
   * @type {string}
   * @memberof Question
   */
  title?: string;
  /**
   *
   * @type {QuestionTypeEnum}
   * @memberof Question
   */
  type?: QuestionTypeEnum;
}

/**
 * @export
 */
export const QuestionTypeEnum = {
  Text: 'TEXT',
  Number: 'NUMBER',
  Date: 'DATE',
  SingleSelect: 'SINGLE_SELECT',
  MultiSelect: 'MULTI_SELECT',
} as const;
export type QuestionTypeEnum = (typeof QuestionTypeEnum)[keyof typeof QuestionTypeEnum];

/**
 * Check if a given object implements the Question interface.
 */
export function instanceOfQuestion(value: object): value is Question {
  return true;
}

export function QuestionFromJSON(json: any): Question {
  return QuestionFromJSONTyped(json, false);
}

export function QuestionFromJSONTyped(json: any, ignoreDiscriminator: boolean): Question {
  if (json == null) {
    return json;
  }
  return {
    code: json['code'] == null ? undefined : json['code'],
    required: json['required'] == null ? undefined : json['required'],
    section: json['section'] == null ? undefined : json['section'],
    title: json['title'] == null ? undefined : json['title'],
    type: json['type'] == null ? undefined : json['type'],
  };
}

export function QuestionToJSON(json: any): Question {
  return QuestionToJSONTyped(json, false);
}

export function QuestionToJSONTyped(
  value?: Question | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    code: value['code'],
    required: value['required'],
    section: value['section'],
    title: value['title'],
    type: value['type'],
  };
}